requirements.txt
Flask
Flask-SQLAlchemy
flask-session
bcrypt
gunicorn
python-dotenv
mysql-connector-python
pytest
Flask-JWT-Extended